+++ History.md
@@ -1,7 +1,8 @@
-1.0.8 / 2014-07-??
+1.0.8 / 2014-07-24
 ==================
   * JSCover 1.0.13
   * Add logging
+  * Improve robustness
   * Internal: Extract common test runner code
 
 1.0.7 / 2014-07-03